apigw-rest-api/pyproject.toml
[project]
name = "hello-api"
version = "0.1.0"
description = "A Hello World FastAPI application with OpenTelemetry instrumentation"
readme = "README.md"
authors = [
    { name = "phitoduck", email = "eric.riddoch@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi-slim",
    "pydantic>=2",
    "pydantic-settings",
    "structlog",
    "httpx",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-http",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-logging",
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov"]
asgi = ["uvicorn"]
lambda = ["mangum"]
dev = ["hello-api[test,asgi,lambda]"]

[dependency-groups]
aws-lambda = [
    "mangum",
]

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"
